internal/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
h11==0.14.0
idna==3.10
nats-py==2.9.0
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
redis==5.2.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
websockets==14.1
